Update dataPreprocessingBE.py
Heat recovery ratio "Wirkungsgrad der Wärmerückgewinnungseinheit RLT (heat_recovery_efficiency)" was adapted to the final dataset of DataNWG BRE. Therefore, qH3 was replaced with qh3_1. This Variable contains the information if heat recovery is implemented or not. The encoding was simplified . The old Version is still available uncommented below.