dataPreprocessingBE.py improvment and documentation of changes in seperat txt file
Several improvements were added to the dataPreprocessingBE.py, as a detailed verification of the initially made assumptions and equations was undertaken, revealing some typo bugs that were removed. One Bug was the missing consideration of n_UG in calculating the base_area. This influences the A/V and therefore
to a certain amount, the energy balance (generally removing slightly the space heating energy demand).