Improvement of lighting energy calculation
Detailed analysis of the DIBS results applied to the German non-domestic building stock showed that the lighting demand was overestimated. The reason was the assumption that the entire building area is lighted in the case of any occupancy. This has been changed so that the occupancy percentage defines the area portion of the building lighted. Thus, the overestimation of the lighting demand has been corrected.