Negative appliance gains in building for food was fixed. Further small improvments on calculation time were implemented.
Appliance_gains equal the electric energy that appliances use, except for negative appliance_gains of refrigerated counters in trade buildings for food. This case is now specially considered. In this context Appliance_gains_elt_demand_sum for the electrical share of the Appliance_gains_demand_sum was established. Additionally the individual saving of each building was removed. From no on only all buildings are saved once at the end of the DIBS calculation, to save caluclation time.