Update annualSimulation.py
Corrected Bug in writing hourly results of Appliance_gains_demands and Appliance_gains_elt_demands. The calculation of both was and is correct in DIBS, but the values were saved incorrectly, resulting in the last-hour value being written in all hours and summed up to the annual values. Consequently, the Appliance_gains_demand_sum and Appliance_gains_elt_demand_sum were wrong in the Excel outputs. This is fixed now. The Bug did not affect any other output results, as the calculation was and is performing correctly.